requirements.txt
# NexSupply AI - Python Dependencies
# Last updated: 2025-01-XX

# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0

# AI/ML
google-generativeai>=0.3.0

# Environment & Config
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2

# Optional: PDF Generation (if needed)
# fpdf2>=2.7.0

# Development (optional)
# pytest>=7.4.0
# black>=23.0.0
# ruff>=0.1.0
